fix: translations, download links, os detection
